refactor: step2(피드백) - 패키지 구조 통일
