+++ src/main/java/nextstep/subway/line/dto/LineRequest.java
@@ -1,8 +1,7 @@
-package nextstep.subway.dto;
+package nextstep.subway.line.dto;
 
-import java.util.Objects;
-import nextstep.subway.domain.Line;
-import nextstep.subway.domain.Station;
+import nextstep.subway.line.domain.Line;
+import nextstep.subway.station.domain.Station;
 
 public class LineRequest {
