+++ src/main/java/nextstep/subway/station/service/StationService.java
@@ -1,11 +1,11 @@
-package nextstep.subway.application;
+package nextstep.subway.station.service;
 
 import java.util.List;
 import java.util.stream.Collectors;
-import nextstep.subway.domain.Station;
-import nextstep.subway.domain.StationRepository;
-import nextstep.subway.dto.StationRequest;
-import nextstep.subway.dto.StationResponse;
+import nextstep.subway.station.domain.Station;
+import nextstep.subway.station.repository.StationRepository;
+import nextstep.subway.station.dto.StationRequest;
+import nextstep.subway.station.dto.StationResponse;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
