+++ src/test/java/nextstep/subway/line/LineAcceptanceTest.java
@@ -10,9 +10,9 @@ import io.restassured.response.ValidatableResponse;
 import java.util.Arrays;
 import java.util.List;
 import nextstep.subway.BaseAcceptanceTest;
-import nextstep.subway.dto.LineRequest;
-import nextstep.subway.dto.LineResponse;
-import nextstep.subway.dto.StationResponse;
+import nextstep.subway.line.dto.LineRequest;
+import nextstep.subway.line.dto.LineResponse;
+import nextstep.subway.station.dto.StationResponse;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
